lalalala
- CMSG_LOGOUT_REQUEST  *Partial
- CMSG_LOGOUT_CANCEL *DONE
- Added Move Root UnRoot
- CMSG_STANDSTATECHANGE
- Removed not used code
- EmotesText.DBC Added
- CMSG_TEXT_EMOTE improvements

+++ Common/Globals/Global.Opcodes.cs
@@ -1,7 +1,7 @@
 ﻿namespace Common.Globals
 {
 
-    public enum StandStates : byte
+    public enum StandStates : int
     {
         STANDSTATE_STAND = 0,
         STANDSTATE_SIT = 1,
